50 first fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-int i, a, b, c;
+long int i, a, b, c;
 a = 1;
 b = 2;
 printf("%d, %d, ", a, b);
